Increase default wait time to 5s
Occassionally we see failures at random places under macos or ARM64 from
not waiting long enough.

Increasing the default to 5s ought to help with these, and shouldn't
slow anything down that doesn't need it (since this is just a max wait
time before waiting).

This also fixes a (required) timeout by changing it to 500ms so that the
test suite doesn't take longer to run as a result of the above change.