Properly handle dual-stack IPv6
Default-constructed Address objects (or objects with an empty string IP)
were becoming IPv6 any address (`::`).

This doesn't work everywhere: on Windows, some Linux, and most BSDs
sockets are not dual stack by default, and so binding to this address
could only talk to IPv6 targets but not IPv4 targets.

For most platforms we can fix that by explicitly setting the IPV6_V6ONLY
flag to 0 on the socket during socket initialization.  (OpenBSD and
DragonFly BSD, however, don't allow this.)

Thus this fix:

- the Address object gains a `dual_stack` flag, which is turned on only
  for default-constructed/empty-host Address objects.
  - NB: this means Address{"::", 0} and Address{"", 0} are different
    (only the second will be dual_stack).
- on OpenBSD and DragonFly a default constructed Address now becomes an
  IPv4 any addr (0.0.0.0) because our only options are only-IPv6 or
  only-IPv4, and in such a case IPv4 is a much better default.
- everywhere else, default/empty construction now becomes IPv6 `::`
  *and* we flip on the dual_stack flag.
- when binding an IPv6 socket we now force IPV6_V6ONLY on/off based on
  whether the Address's `dual_stack` is set off/on, so that our
  behaviour will be consistent on all platforms (except for
  OpenBSD/DragonFLy, which simply can't be dual stack).