clang-format: v16, c++20, remove semicolons, force west const
Removing superfluous semicolons (requires 16+) mostly affects the test
suite where we have a bunch of semicolons after SECTIONs and TEST_CASEs.

Forcing west const didn't change anything but is useful when copying
code in from elsewhere that uses east const.

Changing to c++20 matches the code we use, but also is needed to
properly format the concept in formattable.hpp without butchering it or
trying to remove the semicolon.