Fix race condition in dual-connection set_close_quietly test
The key verification callback here can rarely fire *before* the
assignment to `server_ci` and `client_ci` complete, which will segfault
because the `server/client_ci` won't be set yet for the lambda to
invoke.

This fixes it with a mutex to delay the body of the key verification
callbacks until assignment to the two `..._ci` variables is complete.