Allowing specifying a dgram data callback in listen/connect
It felt restrictive (and a little unintuitive) to have to specify the
dgram callback at the endpoint because then I can't have a different
callback for individual connections or for incoming versus outgoing
connections.  It also feels a little weird because stream callbacks are
only at the listen/connect level but not at the endpoint level.

This commit allows also specifying the datagram callback to `connect()`
or `listen()` -- if specified as such, it takes precedence over the one
given to endpoint.