BTRequestStream: fire callbacks if invoked on closing stream
There is a race condition possible where a connection closes before the
application notices, but meanwhile the application submits a command.

If the connection close happens before the command, then the close
triggers all the stream close callbacks, but then the subsequent
`command()` gets added to `sent_reqs` but never fired (since we only
fire closed callbacks once).

This addresses it by firing the callback (immediately) if a command with
a callback gets invoked on a closed connection, just as would have
happened if the `command()` arrived just *before* the connection closed.

This also makes some other small changes closely related to this:

- rearranges the BTRequestStream closed() call to use the superclass
  method to invoke the close_callback, rather than doing it itself.
- moved the try/catch around the stream close callback from
  BTRequestStream into Stream (which didn't have one).
- also set `_is_shutdown` and `_is_closing` to true in the
  Stream::closed so that `is_closing()` can be used to determine that
  the Stream was closed.  Previously these were only set for
  specifically closed streams, but not streams closed implicitly by a
  connection closing.