Dead conn stream handling redux
The previous commit exposed a clunky interface to applications: rather
than simply dealing with failed streams in a close callback, they would
have to deal with it there *and* in a try/catch around stream
construction, and resulted in different behaviours depending on which
way the race resolved.

This redoes it: in a stream creation / connection closing race, we now
always close the stream, regardless of the race outcome.  If the
connection closes first, then we still create the stream but then
immediately fire its close callback.

This allows application code to not have to worry about exceptions
during stream construction and having two different paths to cleanup:
they can simply rely on a stream closed callback getting closed
regardless of the timing of a racing connection close.