Remove sent_request::payload(), ::view()
The rvalue qualified payload() doesn't really work anymore with the fix
in the previous commit where we need to transfer the payload *without*
losing the object, and since `data` is public already, it seems easier
to just use it directly and drop the rvalue-only `payload()` method.

Also removed `view()` because, with the previous fix, it isn't used
anywhere and someone who wants a view can just get it implicitly from
the public `data` member which is all it viewed anyway.