Remove [square-brackets] from Address::host()
Having square brackets in the return makes little sense: it is a
convention used only when passing a combined host+port value, and causes
difficulty when trying to use it with code that doesn't use that
convention (e.g. by passing ip and port separately).

to_string() should indeed have the square brackets since it returns a
combined value, and keeps it in this change.